requirements.txt
streamlit==1.33.0
datetime
python-dotenv==1.0.1
google.generativeai==0.7.2
streamlit_chat==0.1.1 
PIL